requirements.txt
requests==2.31.0
pandas==2.0.3
sqlalchemy==2.0.20
psycopg2-binary==2.9.9
python-dotenv==1.0.0
urllib3==2.0.4
